pyproject.toml
[tool.poetry]
name = "hopfield"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.7"
numpy = "^1.18.1"
Pillow = "^7.0.0"

[tool.poetry.dev-dependencies]
pytest = "^4.6"

[build-system]
requires = ["poetry>=1.0.0"]
build-backend = "poetry.masonry.api"
